# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super STT daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/stt/super-stt-http.sock` (override with `SUPER_STT_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/stt/super-stt-http.sock\" \\\n     -H \"Authorization: Bearer $STT_TOKEN\" \\\n     http://stt.local/v1/ping\n```",
    "contact": {
      "name": "Super STT",
      "url": "https://github.com/jorge-menjivar/super-stt"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.2.4-beta.1"
  },
  "servers": [
    {
      "url": "http://stt.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/language": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Read the default transcription language",
        "description": "A BCP-47 tag, `auto`, or `null` when nothing is configured.",
        "operationId": "get_language",
        "responses": {
          "200": {
            "description": "Done.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageState"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Set the default transcription language",
        "description": "Sets the language every model transcribes in unless something more specific overrides it: a per-model setting, or a `language` field in a single `POST /transcribe` body.",
        "operationId": "set_language",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLanguageBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageState"
                }
              }
            }
          },
          "400": {
            "description": "The value was rejected — out of range, or not one of the accepted tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Clear the default transcription language",
        "description": "Removes the global setting, returning every model to detecting the language itself. Per-model overrides are untouched.",
        "operationId": "clear_language",
        "responses": {
          "200": {
            "description": "Done.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageState"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/transcribe": {
      "post": {
        "tags": [
          "transcribe"
        ],
        "summary": "Transcribe supplied audio, or start a microphone recording",
        "description": "One endpoint with four behaviours, selected by the body:\n\n| Body | Behaviour | Response |\n|---|---|---|\n| `audio_data` present | Transcribe the supplied buffer; the microphone is never opened | `200` JSON with `transcription` |\n| no `audio_data`, `wait: false` (default) | Start recording and detach; stop it with `POST /transcribe/stop` | `202` with `message: \"Recording started\"` |\n| `wait: true` | Record, then stream the result | `200 text/event-stream`, one `done` frame |\n| `wait: true`, `stream_realtime: true` | As above, plus incremental previews | `200 text/event-stream`, `preview` frames then `done` |\n\nOn the streaming paths a daemon-side failure arrives as a single `error` frame, and closing the connection stops the recording. Frames are:\n\n| `event:` | `data:` |\n|---|---|\n| `preview` | `{ \"text\": \"hello wor…\" }` |\n| `done` | `{ \"transcription\": \"hello world\" }` |\n| `error` | `{ \"message\": \"…\" }` |\n\nStarting while a recording is already in flight is `409 recording_in_progress` — this endpoint only ever *starts* one. Read `busy` from `GET /status` and call `POST /transcribe/stop` to end the running capture.",
        "operationId": "transcribe",
        "requestBody": {
          "description": "All fields optional; the combination selects the behaviour.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Pre-captured audio: the finished transcription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transcription"
                }
              }
            }
          },
          "202": {
            "description": "Microphone recording started and detached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                },
                "example": {
                  "message": "Recording started",
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "`audio_data` was not an array of numbers, or `stream_realtime` was sent with it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `transcribe` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "A recording is already in flight (`recording_in_progress`), or no model is loaded (`model_not_loaded`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "transcribe"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ack": {
        "type": "object",
        "description": "The plain acknowledgement: an operation succeeded, with a sentence saying\nwhat happened.\n\nA good number of settings endpoints answer with exactly this, on `GET` as\nwell as `POST` — `GET /volume` reports the level inside `message` rather\nthan as a field of its own. Where that is true the endpoint's own\ndocumentation says so, because a client has to parse the number back out.",
        "required": [
          "status"
        ],
        "properties": {
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail. Not a stable identifier — do not switch on it.\n\nOptional because a few commands acknowledge without a sentence, and the\nkey is then absent rather than empty. Every endpoint documented as\ncarrying its value in the message always sets it."
          },
          "status": {
            "type": "string",
            "description": "Always `success`.",
            "example": "success"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "recording_in_progress",
          "download_in_progress",
          "no_switch_in_progress",
          "model_not_loaded",
          "invalid_model",
          "invalid_backend",
          "cuda_unavailable",
          "invalid_device",
          "invalid_audio_theme",
          "unsupported_language",
          "invalid_value",
          "not_found",
          "internal",
          "unknown"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-stt/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "LanguageState": {
        "type": "object",
        "description": "The default transcription language.",
        "required": [
          "status"
        ],
        "properties": {
          "language": {
            "type": [
              "string",
              "null"
            ],
            "description": "A BCP-47 tag, `auto`, or `null` when nothing is configured. Always\npresent — `null` is the answer, not an absent key.",
            "example": "es"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "SetLanguageBody": {
        "type": "object",
        "description": "Set the default transcription language",
        "required": [
          "language"
        ],
        "properties": {
          "language": {
            "type": "string",
            "description": "A BCP-47 tag such as `es`, or `auto` to let the model detect the language."
          }
        }
      },
...
      "TranscribeBody": {
        "type": "object",
        "description": "Pre-captured one-shot: transcribe a supplied `audio_data` buffer without\ntouching the microphone and return a single JSON `{ transcription }` (or a\ncoded error). Rejects `stream_realtime` combined with `audio_data` per the\ncontract.\nThe body of `POST /transcribe`. Every field is optional; which combination\nyou send selects one of the four behaviours the endpoint documents.\n\nDeclared, not deserialized. The handler reads the body as raw JSON so it can\n*move* a large `audio_data` buffer out of it rather than copying it into a\nsecond owned `Vec`; this type is what that body's shape is, published in the\n`OpenAPI` document and kept beside the handler that parses it.",
        "properties": {
          "audio_data": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Mono PCM samples in `-1.0..=1.0`. Present means \"transcribe this buffer\"\n— the daemon does not touch the microphone, and answers with one JSON\nresult rather than a stream."
          },
          "language": {
            "type": [
              "string",
              "null"
            ],
            "description": "BCP-47 tag, or `auto`, overriding the configured language for this\nrequest only.",
            "example": "es"
          },
          "sample_rate": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Sample rate of `audio_data`, in Hz. Ignored on the microphone paths.",
            "example": 16000,
            "minimum": 0
          },
          "stream_realtime": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Microphone paths only, and only with `wait: true`: also emit incremental\n`preview` frames as the transcript forms. Rejected with `400` alongside\n`audio_data`, which has nothing to stream."
          },
          "wait": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Microphone paths only. `false` (the default) starts the recording and\nreturns immediately; `true` holds the connection open and streams the\nresult."
          }
        }
      },
      "Transcription": {
        "type": "object",
        "description": "A one-shot transcription result, for the pre-captured path.",
        "required": [
          "status",
          "transcription"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "Always `success`.",
            "example": "success"
          },
          "transcription": {
            "type": "string",
            "description": "The recognized text. Empty when the audio held no speech — which is a\nsuccessful transcription, not a failure.",
            "example": "hello world"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}